Crate twothousand_forty_eight

a basic engine and move validator for the game 2048

Includes wasm functions generated with wasm_bindgen

Modules

board

Provides Board to hold game board data and tile to hold the values of the board tiles.

direction

Provides Direction to represent move directions

parser

recording

Provides Recording, to hold recorded games

validator

Provides functions to validate a Recording

Constants

DEBUG_INFO

Functions

add_randomwasm

add_random_to_boardadd_random

apply_movewasm

get_frameswasm

get_random_tile_to_addadd_random

hashwasm and history_hash

parsewasm

validatewasm