Module twothousand_forty_eight::board

Provides Board to hold game board data and tile to hold the values of the board tiles.

Modules

tile

Structs

Board

Holds game board data

Constants

MAX_HEIGHT

Max height of a board the program can handle. Be careful when increasing, as this increases memory use expotentially.

MAX_WIDTH

Max width of a board the program can handle. Be careful when increasing, as this increases memory use expotentially.

Functions

board_to_string

Return a debug visualization of the board

create_tiles

Initialize an array of empty tiles created with Tile::new

get_closest_tile

Return the closest tile with the value of “mask” to the tile “t” in the given direction “dir”, if “t” is returned, no such tile was found.

get_farthest_tile

Return the farthest tile with the value of “mask” to the tile “t” in the given direction “dir”, if “t” is returned, no such tile was found.

is_move_possible

Check if a move is possible in the direction “dir” and return the next board, the possibility, and the score gain

print_board

Print a debug visualization of the board