Function twothousand_forty_eight::parser::parse_data

pub fn parse_data(data: String) -> Recording

Parses a string representation of a played 2048 game into a Recording

The string should be in the following format:

The string can start with [w]x[h]S, specifying the size of the game board, otherwise it defaults to 4 by 4. (w = width, h = height, both are a usize)

History indicies should be separated by a single :

; separates the history index to:

The board data of the move on the left side, separated by a +:

Each tile value of the board, separated from eachother by dots (.) on the left side

The tile added to the board after this move on the right side

The tile’s value is on the right side of a comma (,) and it’s position is on the left side of the comma, separated by a dot (.), x first

Direction index of the move on the right side

e.g. 4x4S0.0.0.0.0.0.0.0.0.0.0.0.0.0.2.2+2,1.2;1