Module twothousand_forty_eight::validator

Provides functions to validate a Recording

Functions

get_run_score

Returns the accumulated score of a run (should match the score displayed in the game)

validate_first_move

Makes sure that the starting state of the history doesn’t contain too many tiles or tiles with a too big value.

validate_history

Validates the history continuity and returns the determined validity, score, possible score margin (caused when the last game move was not recorded) and break count.